01_docker_compose/app/requirements.txt
Django==3.2.6
psycopg2-binary==2.9.1
gunicorn
django-split-settings
python-dotenv
django-environ==0.8.1
djangorestframework
markdown
django-filter
drf-yasg